Dev

See merge request docudots/banking-admin-backend!14

+++ api/database/migrations/2021_09_16_104537_create_group_role_table.php
@@ -15,8 +15,8 @@ class CreateGroupRoleTable extends Migration
     {
         Schema::create('group_role', function (Blueprint $table) {
             $table->id();
-            $table->integer('group_id');
-            $table->integer('role_id');
+            $table->unsignedBigInteger('group_id');
+            $table->unsignedBigInteger('role_id');
             $table->foreign('group_id')->references('id')->on('groups');
             $table->foreign('role_id')->references('id')->on('roles');
         });
